+++ src/app/draggable-example/draggable-example.component.spec.ts
@@ -1,4 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { MatCardModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipDraggableModule } from 'pip-webui2-behaviors';
 
 import { DraggableExampleComponent } from './draggable-example.component';
 
@@ -8,9 +12,16 @@ describe('DraggableExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ DraggableExampleComponent ]
+      declarations: [DraggableExampleComponent],
+      imports: [
+        CommonModule,
+        MatCardModule,
+        TranslateModule.forRoot(),
+
+        PipDraggableModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {