+++ src/app/focused-example/focused-example.component.spec.ts
@@ -1,4 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatListModule, MatToolbarModule, MatCardModule, MatIconModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipFocusedModule } from 'pip-webui2-behaviors';
 
 import { FocusedExampleComponent } from './focused-example.component';
 
@@ -8,9 +13,20 @@ describe('FocusedExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ FocusedExampleComponent ]
+      declarations: [FocusedExampleComponent],
+      imports: [
+        CommonModule,
+        FlexLayoutModule,
+        MatListModule,
+        MatToolbarModule,
+        MatCardModule,
+        MatIconModule,
+        TranslateModule.forRoot(),
+
+        PipFocusedModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {